library-cleanup: Don't build libNX_Xau anymore. Use system's libXau shared library.
