--- debian/libnx-xau-dev.install
@@ -1,2 +1,0 @@
-usr/lib/*/libNX_Xau.so
-usr/include/*/nx-X11/Xauth.h
