--- debian/libnx-xau6.install
@@ -1,1 +1,0 @@
-usr/lib/*/libNX_Xau.so.*
