+++ nx-X11/config/cf/README
@@ -189,7 +189,6 @@ X11.tmpl provides defaults for the following variables:
 	ConnectionFlags		-D's for supported transports
 	ContribDir		location of user-contributed sources
 	DebugLibX11		build libX11_d.a
-	DebugLibXau		build libXau_d.a
 	DebugLibXext		build libXext_d.a
 	DefaultCursorTheme      name of default cursor theme
 	DefaultFontPath		default server font path
@@ -223,12 +222,10 @@ X11.tmpl provides defaults for the following variables:
 	NeedDefaultDepLibs	boolean for enabling default DEPLIBS
 	NlsDir			directory in which to install nls files
 	NormalLibX11		build libX11.a
-	NormalLibXau		build libXau.a
 	NormalLibXext		build libXext.a
 	OsNameDefines		If uname(2) unavailable, set to -DOS_NAME=OSName
 	PrimaryScreenResolution	resolution of default server screen
 	ProfileLibX11		build libX11_p.a
-	ProfileLibXau		build libXau_p.a
 	ProfileLibXext		build libXext_p.a
 	ProjectX		version indicating this is the X Window System
 	RemoveTargetProgramByMoving	boolean for rm -f that doesn't
@@ -239,7 +236,6 @@ X11.tmpl provides defaults for the following variables:
 	ServerOSDefines		OS -D's for server
 	ServerAssertDefines	-DNDEBUG for no assertions, /**/ for assertions
 	SharedLibX11		boolean for making sharable libX11.so
-	SharedLibXau		boolean for making sharable libXau.so
 	SharedLibXext		boolean for making sharable libXext.so
 	ShmDefines		MIT-SHM define
 	TwmDir			directory in which to install twm config files
@@ -311,8 +307,6 @@ An <os>Lib.tmpl file provides defaults for the following variables:
 
 	SharedX11Reqs		required libs for libX11.so
 	SharedX11Rev		version number for libX11.so
-	SharedXauReqs		required libs for libXau.so
-	SharedXauRev		version number for libXau.so
 	SharedXextReqs		required libs for libXext.so
 	SharedXextRev		version number for libXext.so
