+++ nx-X11/lib/X11/AuDispose.c
