+++ nx-X11/lib/X11/AuGetBest.c
@@ -1,4 +1,4 @@
-/* $Xorg: AuGetAddr.c,v 1.4 2001/02/09 02:03:42 xorgcvs Exp $ */
+/* $Xorg: AuGetBest.c,v 1.4 2001/02/09 02:03:42 xorgcvs Exp $ */
 
 /*
 
@@ -25,13 +25,21 @@ used in advertising or otherwise to promote the sale, use or other dealings
 in this Software without prior written authorization from The Open Group.
 
 */
-/* $XFree86: xc/lib/Xau/AuGetAddr.c,v 1.5 2001/12/14 19:54:36 dawes Exp $ */
+/* $XFree86: xc/lib/Xau/AuGetBest.c,v 1.7 2001/12/14 19:54:36 dawes Exp $ */
 
 #ifdef HAVE_CONFIG_H
 #include <config.h>
 #endif
 #include <nx-X11/Xauth.h>
 #include <nx-X11/Xos.h>
+#ifdef XTHREADS
+#include <nx-X11/Xthreads.h>
+#endif
+#ifdef hpux
+#define X_INCLUDE_NETDB_H
+#define XOS_USE_NO_LOCKING
+#include <nx-X11/Xos_r.h>
+#endif
 
 static int
 binaryEqual (_Xconst char *a, _Xconst char *b, int len)
@@ -43,31 +51,35 @@ binaryEqual (_Xconst char *a, _Xconst char *b, int len)
 }
 
 Xauth *
-XauGetAuthByAddr (
+XauGetBestAuthByAddr (
 #if NeedWidePrototypes
-unsigned int	family,
-unsigned int	address_length,
+    unsigned int	family,
+    unsigned int	address_length,
 #else
-unsigned short	family,
-unsigned short	address_length,
+    unsigned short	family,
+    unsigned short	address_length,
 #endif
-_Xconst char*	address,
+    _Xconst char*	address,
 #if NeedWidePrototypes
-unsigned int	number_length,
+    unsigned int	number_length,
 #else
-unsigned short	number_length,
+    unsigned short	number_length,
 #endif
-_Xconst char*	number,
-#if NeedWidePrototypes
-unsigned int	name_length,
-#else
-unsigned short	name_length,
-#endif
-_Xconst char*	name)
+    _Xconst char*	number,
+    int			types_length,
+    char**		types,
+    _Xconst int*	type_lengths)
 {
     FILE    *auth_file;
     char    *auth_name;
     Xauth   *entry;
+    Xauth   *best;
+    int	    best_type;
+    int	    type;
+#ifdef hpux
+    char		*fully_qual_address;
+    unsigned short	fully_qual_address_length;
+#endif
 
     auth_name = XauFileName ();
     if (!auth_name)
@@ -77,6 +89,29 @@ _Xconst char*	name)
     auth_file = fopen (auth_name, "rb");
     if (!auth_file)
 	return 0;
+
+#ifdef hpux
+    if (family == FamilyLocal) {
+#ifdef XTHREADS_NEEDS_BYNAMEPARAMS
+	_Xgethostbynameparams hparams;
+#endif
+	struct hostent *hostp;
+
+	/* make sure we try fully-qualified hostname */
+	if ((hostp = _XGethostbyname(address,hparams)) != NULL) {
+	    fully_qual_address = hostp->h_name;
+	    fully_qual_address_length = strlen(fully_qual_address);
+	}
+	else
+	{
+	    fully_qual_address = NULL;
+	    fully_qual_address_length = 0;
+	}
+    }
+#endif /* hpux */
+
+    best = 0;
+    best_type = types_length;
     for (;;) {
 	entry = XauReadAuth (auth_file);
 	if (!entry)
@@ -96,17 +131,43 @@ _Xconst char*	name)
 
 	if ((family == FamilyWild || entry->family == FamilyWild ||
 	     (entry->family == family &&
-	      address_length == entry->address_length &&
-	      binaryEqual (entry->address, address, (int)address_length))) &&
+	     ((address_length == entry->address_length &&
+	     binaryEqual (entry->address, address, (int)address_length))
+#ifdef hpux
+	     || (family == FamilyLocal &&
+		fully_qual_address_length == entry->address_length &&
+	     	binaryEqual (entry->address, fully_qual_address,
+		    (int) fully_qual_address_length))
+#endif
+	    ))) &&
 	    (number_length == 0 || entry->number_length == 0 ||
 	     (number_length == entry->number_length &&
-	      binaryEqual (entry->number, number, (int)number_length))) &&
-	    (name_length == 0 || entry->name_length == 0 ||
-	     (entry->name_length == name_length &&
- 	      binaryEqual (entry->name, name, (int)name_length))))
-	    break;
+	      binaryEqual (entry->number, number, (int)number_length))))
+	{
+	    if (best_type == 0)
+	    {
+		best = entry;
+		break;
+	    }
+	    for (type = 0; type < best_type; type++)
+		if (type_lengths[type] == entry->name_length &&
+		    !(strncmp (types[type], entry->name, entry->name_length)))
+		{
+		    break;
+		}
+	    if (type < best_type)
+	    {
+		if (best)
+		    XauDisposeAuth (best);
+		best = entry;
+		best_type = type;
+		if (type == 0)
+		    break;
+		continue;
+	    }
+	}
 	XauDisposeAuth (entry);
     }
     (void) fclose (auth_file);
-    return entry;
+    return best;
 }
