--- nx-X11/lib/Xau/Autest.c
@@ -1,71 +1,0 @@
-/* $Xorg: Autest.c,v 1.4 2001/02/09 02:03:42 xorgcvs Exp $ */
-
-/*
-
-Copyright 1988, 1998  The Open Group
-
-Permission to use, copy, modify, distribute, and sell this software and its
-documentation for any purpose is hereby granted without fee, provided that
-the above copyright notice appear in all copies and that both that
-copyright notice and this permission notice appear in supporting
-documentation.
-
-The above copyright notice and this permission notice shall be included in
-all copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL THE
-OPEN GROUP BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN
-AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-Except as contained in this notice, the name of The Open Group shall not be
-used in advertising or otherwise to promote the sale, use or other dealings
-in this Software without prior written authorization from The Open Group.
-
-*/
-
-#ifdef HAVE_CONFIG_H
-#include <config.h>
-#endif
-#include <nx-X11/Xauth.h>
-
-main (argc, argv)
-char	**argv;
-{
-    Xauth   test_data;
-    char    *name, *data, *file;
-    int	    state = 0;
-    FILE    *output;
-
-    while (*++argv) {
-	if (!strcmp (*argv, "-file"))
-	    file = *++argv;
-	else if (state == 0) {
-	    name = *argv;
-	    ++state;
-	} else if (state == 1) {
-	    data = *argv;
-	    ++state;
-	}
-    }
-    if(!file) {
-	fprintf (stderr, "No file\n");
-	exit (1);
-    }
-    test_data.family = 0;
-    test_data.address_length = 0;
-    test_data.address = "";
-    test_data.number_length = 0;
-    test_data.number = "";
-    test_data.name_length = strlen (name);
-    test_data.name = name;
-    test_data.data_length = strlen (data);
-    test_data.data = data;
-    output = fopen (file, "w");
-    if (output) {
-	XauWriteAuth (output, &test_data);
-	fclose (output);
-    }
-}
