--- nx-X11/lib/Xau/Imakefile
@@ -1,49 +1,0 @@
-XCOMM $Xorg: Imakefile,v 1.3 2000/08/17 19:45:29 cpqbld Exp $
-XCOMM $XdotOrg: xc/lib/Xau/Imakefile,v 1.4 2005/09/22 23:54:18 alanc Exp $
-
-
-
-XCOMM $XFree86: xc/lib/Xau/Imakefile,v 3.5 1999/04/17 09:08:11 dawes Exp $
-
-#define DoNormalLib NormalLibXau
-#define DoSharedLib SharedLibXau
-#define DoExtraLib SharedLibXau
-#define DoDebugLib DebugLibXau
-#define DoProfileLib ProfileLibXau
-#define LibName NX_Xau
-#define SoRev SOXAUTHREV
-#define IncSubdir nx-X11
-
-#include <Threads.tmpl>
-
-#ifdef SharedXauReqs
-REQUIREDLIBS = SharedXauReqs
-#endif
-
-       LINTLIBS = $(LINTXLIB)
-
-#if HasKrb5
-K5ENCSRC = k5encode.c
-K5ENCOBJ = k5encode.o
-INCLUDES = Krb5Includes
-#else
-K5ENCSRC =
-K5ENCOBJ =
-INCLUDES =
-#endif
-
-HEADERS = \
-	Xauth.h
-
-SRCS =	AuDispose.c AuFileName.c AuGetAddr.c AuGetBest.c AuLock.c \
-	AuRead.c AuUnlock.c AuWrite.c $(K5ENCSRC)
-
-OBJS =	AuDispose.o AuFileName.o AuGetAddr.o AuGetBest.o AuLock.o \
-	AuRead.o AuUnlock.o AuWrite.o $(K5ENCOBJ)
-
-#define IncludeSharedObjectInNormalLib
-#define UnsharedLibObjCompile(options) LibObjCompile(unshared,$(CDEBUGFLAGS) $(CLIBDEBUGFLAGS) options $(PICFLAGS))
-
-#include <Library.tmpl>
-
-DependTarget()
